+++ packages/common-ts/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Upgrade dependencies
+
+### Fixed
+- Fix serve metrics test
 
 ## [1.5.1] - 2021-05-11
 ### Added